Refactor AST conversions and other cleanups
