+++ Sources/_StringProcessing/ByteCodeGen.swift
@@ -5,9 +5,10 @@ extension Compiler {
     var options: MatchingOptions
     var builder = _MatchingEngine.Program<String>.Builder()
 
-    mutating func finish() -> _MatchingEngine.Program<String> {
+    mutating func finish(
+    ) throws -> _MatchingEngine.Program<String> {
       builder.buildAccept()
-      return builder.assemble()
+      return try builder.assemble()
     }
   }
 }
@@ -90,7 +91,7 @@ extension Compiler.ByteCodeGen {
 
     case .resetStartOfMatch:
       // FIXME: Figure out how to communicate this out
-      throw unsupported(#"\K (reset/keep assertion)"#)
+      throw Unsupported(#"\K (reset/keep assertion)"#)
 
     case .firstMatchingPositionInSubject:
       // TODO: We can probably build a nice model with API here
@@ -100,11 +101,11 @@ extension Compiler.ByteCodeGen {
 
     case .textSegment:
       // This we should be able to do!
-      throw unsupported(#"\y (text segment)"#)
+      throw Unsupported(#"\y (text segment)"#)
 
     case .notTextSegment:
       // This we should be able to do!
-      throw unsupported(#"\Y (not text segment)"#)
+      throw Unsupported(#"\Y (not text segment)"#)
 
     case .startOfLine:
       builder.buildAssert { (input, pos, bounds) in
@@ -204,7 +205,7 @@ extension Compiler.ByteCodeGen {
     _ child: DSLTree.Node
   ) throws {
     guard kind.forwards else {
-      throw unsupported("backwards assertions")
+      throw Unsupported("backwards assertions")
     }
 
     let positive = kind.positive
@@ -259,7 +260,7 @@ extension Compiler.ByteCodeGen {
     switch kind {
     case .lookahead, .negativeLookahead,
         .lookbehind, .negativeLookbehind:
-      fatalError("unreachable")
+      throw Unreachable("TODO: reason")
 
     case .capture, .namedCapture:
       let cap = builder.makeCapture()
@@ -299,7 +300,7 @@ extension Compiler.ByteCodeGen {
       // Ok
       break
     default:
-      fatalError("unreachable?")
+      throw Unreachable("TODO: reason")
     }
 
     // Compiler and/or parser should enforce these invariants
