+++ Sources/_StringProcessing/ConsumerInterface.swift
@@ -11,58 +11,6 @@
 
 import _MatchingEngine
 
-struct Unsupported: Error, CustomStringConvertible {
-  var message: String
-  var file: String
-  var line: Int
-
-  var description: String { """
-    Unsupported: '\(message)'
-      \(file):\(line)
-    """
-  }
-
-  init(
-    _ s: String,
-    file: StaticString = #file,
-    line: UInt = #line
-  ) {
-    self.message = s
-    self.file = file.description
-    self.line = Int(asserting: line)
-  }
-}
-
-// TODO: Remove
-func unsupported(
-  _ s: String,
-  file: StaticString = #file,
-  line: UInt = #line
-) -> Unsupported {
-  return Unsupported(s, file: file, line: line)
-}
-
-struct Unreachable: Error, CustomStringConvertible {
-  var message: String
-  var file: String
-  var line: Int
-
-  var description: String { """
-    Unreachable: '\(message)'
-      \(file):\(line)
-    """
-  }
-}
-
-func unreachable(
-  _ s: String,
-  file: StaticString = #file,
-  line: Int = #line
-) -> Unreachable {
-  return Unreachable(
-    message: s, file: String(describing: file), line: line)
-}
-
 extension DSLTree.Node {
   /// Attempt to generate a consumer from this AST node
   ///
@@ -233,10 +181,10 @@ extension DSLTree.CustomCharacterClass.Member {
     case let .range(low, high):
       // TODO:
       guard let lhs = low.literalCharacterValue else {
-        throw unsupported("\(low) in range")
+        throw Unsupported("\(low) in range")
       }
       guard let rhs = high.literalCharacterValue else {
-        throw unsupported("\(high) in range")
+        throw Unsupported("\(high) in range")
       }
 
       return { input, bounds in
@@ -315,10 +263,10 @@ extension AST.CustomCharacterClass.Member {
 
     case .range(let r):
       guard let lhs = r.lhs.literalCharacterValue else {
-        throw unsupported("\(r.lhs) in range")
+        throw Unsupported("\(r.lhs) in range")
       }
       guard let rhs = r.rhs.literalCharacterValue else {
-        throw unsupported("\(r.rhs) in range")
+        throw Unsupported("\(r.rhs) in range")
       }
 
       return { input, bounds in
@@ -333,7 +281,7 @@ extension AST.CustomCharacterClass.Member {
 
     case .atom(let atom):
       guard let gen = try atom.generateConsumer(opts) else {
-        throw unsupported("TODO")
+        throw Unsupported("TODO")
       }
       return gen
 
@@ -352,7 +300,8 @@ extension AST.CustomCharacterClass.Member {
       }
 
     case .trivia:
-      throw unreachable("Should have been stripped by caller")
+      throw Unreachable(
+        "Should have been stripped by caller")
 
     case .setOperation(let lhs, let op, let rhs):
       // TODO: We should probably have a component type
@@ -520,22 +469,23 @@ extension AST.Atom.CharacterProperty {
         return value ? cons : invert(cons)
 
       case .script(let s):
-        throw unsupported("TODO: Map script: \(s)")
+        throw Unsupported("TODO: Map script: \(s)")
 
       case .scriptExtension(let s):
-        throw unsupported("TODO: Map script: \(s)")
+        throw Unsupported("TODO: Map script: \(s)")
 
       case .posix(let p):
         return p.generateConsumer(opts)
 
       case .pcreSpecial(let s):
-        throw unsupported("TODO: map PCRE special: \(s)")
+        throw Unsupported("TODO: map PCRE special: \(s)")
 
       case .onigurumaSpecial(let s):
-        throw unsupported("TODO: map Oniguruma special: \(s)")
+        throw Unsupported("TODO: map Oniguruma special: \(s)")
 
       case let .other(key, value):
-        throw unsupported("TODO: map other \(key ?? "")=\(value)")
+        throw Unsupported(
+          "TODO: map other \(key ?? "")=\(value)")
       }
     }()
 
@@ -593,7 +543,8 @@ extension Unicode.BinaryProperty {
       if #available(macOS 10.12.2, iOS 10.2, tvOS 10.1, watchOS 3.1.1, *) {
         return consumeScalarProp(\.isEmojiModifierBase)
       } else {
-        throw unsupported("isEmojiModifierBase on old OSes")
+        throw Unsupported(
+          "isEmojiModifierBase on old OSes")
       }
     case .emojiComponent:
       break
@@ -601,19 +552,20 @@ extension Unicode.BinaryProperty {
       if #available(macOS 10.12.2, iOS 10.2, tvOS 10.1, watchOS 3.1.1, *) {
         return consumeScalarProp(\.isEmojiModifier)
       } else {
-        throw unsupported("isEmojiModifier on old OSes")
+        throw Unsupported("isEmojiModifier on old OSes")
       }
     case .emoji:
       if #available(macOS 10.12.2, iOS 10.2, tvOS 10.1, watchOS 3.1.1, *) {
         return consumeScalarProp(\.isEmoji)
       } else {
-        throw unsupported("isEmoji on old OSes")
+        throw Unsupported("isEmoji on old OSes")
       }
     case .emojiPresentation:
       if #available(macOS 10.12.2, iOS 10.2, tvOS 10.1, watchOS 3.1.1, *) {
         return consumeScalarProp(\.isEmojiPresentation)
       } else {
-        throw unsupported("isEmojiPresentation on old OSes")
+        throw Unsupported(
+          "isEmojiPresentation on old OSes")
       }
     case .extender:
       return consumeScalarProp(\.isExtender)
@@ -701,10 +653,10 @@ extension Unicode.BinaryProperty {
       return consumeScalarProp(\.isXIDStart)
     case .expandsOnNFC, .expandsOnNFD, .expandsOnNFKD,
         .expandsOnNFKC:
-      throw unsupported("Unicode-deprecated: \(self)")
+      throw Unsupported("Unicode-deprecated: \(self)")
     }
 
-    throw unsupported("TODO: map prop \(self)")
+    throw Unsupported("TODO: map prop \(self)")
   }
 }
 
@@ -802,7 +754,8 @@ extension Unicode.ExtendedGeneralCategory {
       ])
 
     case .casedLetter:
-      throw unsupported("TODO: cased letter? not the property?")
+      throw Unsupported(
+        "TODO: cased letter? not the property?")
 
     case .control:
       return consumeScalarGC(.control)
