+++ Sources/_StringProcessing/RegexDSL/Core.swift
@@ -11,15 +11,6 @@
 
 import _MatchingEngine
 
-@dynamicMemberLookup
-public struct RegexMatch<Match> {
-  public let range: Range<String.Index>
-  public let match: Match
-
-  public subscript<T>(dynamicMember keyPath: KeyPath<Match, T>) -> T {
-    match[keyPath: keyPath]
-  }
-}
 
 /// A type that represents a regular expression.
 public protocol RegexProtocol {
@@ -112,83 +103,7 @@ public struct Regex<Match>: RegexProtocol {
   }
 }
 
-extension RegexProtocol {
-  public func match(in input: String) -> RegexMatch<Match>? {
-    _match(
-      input, in: input.startIndex..<input.endIndex)
-  }
-  public func match(in input: Substring) -> RegexMatch<Match>? {
-    _match(
-      input.base, in: input.startIndex..<input.endIndex)
-  }
 
-  // FIXME: This is mostly hacky because we go down two different paths based on
-  // whether there are captures. This will be cleaned up once we deprecate the
-  // legacy virtual machines.
-  func _match(
-    _ input: String,
-    in inputRange: Range<String.Index>,
-    mode: MatchMode = .wholeString
-  ) -> RegexMatch<Match>? {
-    // TODO: Remove this branch when the matching engine supports captures.
-    if regex.hasCapture {
-      let vm = HareVM(program: regex.program.legacyLoweredProgram)
-      guard let (range, captures) = vm.execute(
-        input: input, in: inputRange, mode: mode
-      )?.destructure else {
-        return nil
-      }
-      let convertedMatch: Match
-      if Match.self == (Substring, DynamicCaptures).self {
-        convertedMatch = (input[range], DynamicCaptures(captures)) as! Match
-      } else {
-        let typeErasedMatch = captures.matchValue(
-          withWholeMatch: input[range]
-        )
-        convertedMatch = typeErasedMatch as! Match
-      }
-      return RegexMatch(range: range, match: convertedMatch)
-    }
-
-    let executor = Executor(program: regex.program.loweredProgram)
-    guard let result = executor.execute(
-      input: input, in: inputRange, mode: mode
-    ) else {
-      return nil
-    }
-    let convertedMatch: Match
-    if Match.self == (Substring, DynamicCaptures).self {
-      convertedMatch = (input[result.range], DynamicCaptures.empty) as! Match
-    } else {
-      assert(Match.self == Substring.self)
-      convertedMatch = input[result.range] as! Match
-    }
-    return RegexMatch(range: result.range, match: convertedMatch)
-  }
-}
-
-extension String {
-  public func match<R: RegexProtocol>(_ regex: R) -> RegexMatch<R.Match>? {
-    regex.match(in: self)
-  }
-
-  public func match<R: RegexProtocol>(
-    @RegexBuilder _ content: () -> R
-  ) -> RegexMatch<R.Match>? {
-    match(content())
-  }
-}
-extension Substring {
-  public func match<R: RegexProtocol>(_ regex: R) -> RegexMatch<R.Match>? {
-    regex.match(in: self)
-  }
-
-  public func match<R: RegexProtocol>(
-    @RegexBuilder _ content: () -> R
-  ) -> RegexMatch<R.Match>? {
-    match(content())
-  }
-}
 public struct MockRegexLiteral<Match>: RegexProtocol {
   public typealias MatchValue = Substring
   public let regex: Regex<Match>
